Assignment 1 Part A
This part includes collecting CIKs for each of the Mutual Funds, then looking up the links of all the 13F HRs and the Information Tables, and identifying them either text tables or xml tables. Next, for each text link, extracting the name of issuer, CUSIP, and the Quantity of shares. Finally, writing down the lists in the file.